--- java/src/processing/mode/java/preproc/issue/PdePreprocessIssueException.java
@@ -1,17 +1,0 @@
-package processing.mode.java.preproc.issue;
-
-
-public class PdePreprocessIssueException extends RuntimeException {
-
-  private final PdePreprocessIssue preprocessIssue;
-
-  public PdePreprocessIssueException(PdePreprocessIssue newPreprocessIssue) {
-    super(newPreprocessIssue.getMsg());
-    preprocessIssue = newPreprocessIssue;
-  }
-
-  public PdePreprocessIssue getIssue() {
-    return preprocessIssue;
-  }
-
-}
