+++ java/test/processing/mode/java/ProblemFactoryTest.java
@@ -6,12 +6,10 @@ import org.junit.Test;
 import org.mockito.Mockito;
 import processing.app.Problem;
 import processing.app.ui.Editor;
-import processing.mode.java.preproc.issue.PdePreprocessIssue;
+import processing.mode.java.preproc.PdePreprocessIssue;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import static org.junit.Assert.*;
 
 public class ProblemFactoryTest {
