+++ java/test/processing/mode/java/preproc/AssignmentMessageSimplifierStrategyTest.java
@@ -1,8 +1,10 @@
-package processing.mode.java.preproc.issue;
+package processing.mode.java.preproc;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import processing.mode.java.preproc.PdeIssueEmitter;
+import processing.mode.java.preproc.PreprocessIssueMessageSimplifier;
 
 import java.util.Optional;
 
@@ -18,19 +20,19 @@ public class AssignmentMessageSimplifierStrategyTest {
 
   @Test
   public void testPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("  int x =");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("  int x =");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testPresentDiamond() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("  List<Integer> x =");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("  List<Integer> x =");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testNotPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("class {");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("class {");
     Assert.assertTrue(msg.isEmpty());
   }
