+++ java/test/processing/mode/java/preproc/BadParamMessageSimplifierStrategyTest.java
@@ -1,8 +1,10 @@
-package processing.mode.java.preproc.issue;
+package processing.mode.java.preproc;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import processing.mode.java.preproc.PdeIssueEmitter;
+import processing.mode.java.preproc.PreprocessIssueMessageSimplifier;
 
 import java.util.Optional;
 
@@ -18,25 +20,25 @@ public class BadParamMessageSimplifierStrategyTest {
 
   @Test
   public void testPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("void test (int x,\ny) \n{");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("void test (int x,\ny) \n{");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testPresentUnderscore() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("void test (int x,\ny_y) \n{");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("void test (int x,\ny_y) \n{");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testPresentVarType() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("void test (int x,\nint) \n{");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("void test (int x,\nint) \n{");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testNotPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("int x = y");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("int x = y");
     Assert.assertTrue(msg.isEmpty());
   }
