+++ java/test/processing/mode/java/preproc/ExtraneousInputMessageSimplifierStrategyTest.java
@@ -1,8 +1,10 @@
-package processing.mode.java.preproc.issue;
+package processing.mode.java.preproc;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import processing.mode.java.preproc.PdeIssueEmitter;
+import processing.mode.java.preproc.PreprocessIssueMessageSimplifier;
 
 import java.util.Optional;
 
@@ -19,13 +21,13 @@ public class ExtraneousInputMessageSimplifierStrategyTest {
 
   @Test
   public void testPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("extraneous input 'test' expecting ';'");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("extraneous input 'test' expecting ';'");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testNotPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("String x = \" \\\" \"");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("String x = \" \\\" \"");
     Assert.assertTrue(msg.isEmpty());
   }
