+++ java/test/processing/mode/java/preproc/MessageSimplifierUtilTest.java
@@ -1,8 +1,8 @@
-package processing.mode.java.preproc.issue;
+package processing.mode.java.preproc;
 
 import org.junit.Assert;
 import org.junit.Test;
-import processing.mode.java.preproc.issue.PreprocessIssueMessageSimplifier;
+import processing.mode.java.preproc.PreprocessIssueMessageSimplifier;
 
 public class MessageSimplifierUtilTest {
