+++ java/test/processing/mode/java/preproc/MissingIdentifierMessageSimplifierStrategyTest.java
@@ -1,8 +1,10 @@
-package processing.mode.java.preproc.issue;
+package processing.mode.java.preproc;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import processing.mode.java.preproc.PdeIssueEmitter;
+import processing.mode.java.preproc.PreprocessIssueMessageSimplifier;
 
 import java.util.Optional;
 
@@ -19,13 +21,13 @@ public class MissingIdentifierMessageSimplifierStrategyTest {
 
   @Test
   public void testPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("Missing identifier at ';'");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("Missing identifier at ';'");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testNotPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("String x = \" \\\" \"");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("String x = \" \\\" \"");
     Assert.assertTrue(msg.isEmpty());
   }
