+++ java/test/processing/mode/java/preproc/MissingParenMessageSimplifierStrategyTest.java
@@ -1,9 +1,10 @@
-package processing.mode.java.preproc.issue;
+package processing.mode.java.preproc;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import processing.mode.java.preproc.issue.IssueMessageSimplification;
+import processing.mode.java.preproc.PdeIssueEmitter;
+import processing.mode.java.preproc.PreprocessIssueMessageSimplifier;
 
 import java.util.Optional;
 
@@ -19,13 +20,13 @@ public class MissingParenMessageSimplifierStrategyTest {
 
   @Test
   public void testPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("int x = ((5 + 4) / 3");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("int x = ((5 + 4) / 3");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testNotPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("int x = (y/5)/(\n4)");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("int x = (y/5)/(\n4)");
     Assert.assertTrue(msg.isEmpty());
   }
