+++ java/test/processing/mode/java/preproc/MissingSingleQuoteMessageSimplifierStrategyTest.java
@@ -1,9 +1,10 @@
-package processing.mode.java.preproc.issue;
+package processing.mode.java.preproc;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import processing.mode.java.preproc.issue.IssueMessageSimplification;
+import processing.mode.java.preproc.PdeIssueEmitter;
+import processing.mode.java.preproc.PreprocessIssueMessageSimplifier;
 
 import java.util.Optional;
 
@@ -18,13 +19,13 @@ public class MissingSingleQuoteMessageSimplifierStrategyTest {
 
   @Test
   public void testPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("char x = '");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("char x = '");
     Assert.assertTrue(msg.isPresent());
   }
 
   @Test
   public void testNotPresent() {
-    Optional<IssueMessageSimplification> msg = strategy.simplify("char x = '\\''");
+    Optional<PdeIssueEmitter.IssueMessageSimplification> msg = strategy.simplify("char x = '\\''");
     Assert.assertTrue(msg.isEmpty());
   }
